Change the database interaction routes
Switched the database insert and query rates so we can see that the
database is actually storing data.
Still has problems, though. The original comment said that it picked up
new records and then updated the row to mark it as processed. This is
incorrect. It actually deletes the row after it has been consumed. Need
to do something different to leave the row in the database and mark it
as processed.